+++ back-end/migrations/20201031140205-products.js
@@ -15,7 +15,7 @@ module.exports = {
         unique: true,
       },
       price: {
-        type: Sequelize.DECIMAL(4,2),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       urlImage: {
